:hammer: Extract CodePush functions from publication.js
